feat: add static recommend system to store and manage controller class
