+++ app/src/main/java/javaroke/recommendation/data/saves/original_graph.json
@@ -1,1 +1,1 @@
-{"previousVertex":{"A":{},"B":{},"C":{},"D":{},"E":{},"X":{}},"adjacencyList":{"A":{"A":0.0,"B":6.0,"E":2.0,"X":3.0},"B":{"A":6.0,"B":0.0,"C":2.0},"C":{"B":2.0,"C":0.0,"D":2.0,"X":3.0},"D":{"C":2.0,"D":0.0,"E":2.0},"E":{"A":2.0,"D":2.0,"E":0.0},"X":{"A":3.0,"C":3.0,"X":0.0}},"keyList":["A","B","C","D","E","X"],"size":6,"maximumWeightPosition":{"first":"A","second":"B"},"maximumWeight":6.0,"floorWeight":-1.0}
+{"previousVertex":{"A":{},"B":{},"C":{},"D":{},"E":{},"X":{}},"adjacencyList":{"A":{"A":0.0,"B":2.0,"E":2.0,"X":3.0},"B":{"A":2.0,"B":0.0,"C":2.0},"C":{"B":2.0,"C":0.0,"D":2.0,"X":3.0},"D":{"C":2.0,"D":0.0,"E":2.0},"E":{"A":2.0,"D":2.0,"E":0.0},"X":{"A":3.0,"C":3.0,"X":0.0}},"keyList":["A","B","C","D","E","X"],"size":6,"maximumWeightPosition":{"first":"A","second":"X"},"maximumWeight":3.0,"floorWeight":-1.0}
